backend/pyproject.toml
[project]
name = "Backend"
version = "0.0.1"
description = "Backend for Echo Logic AI Chatbot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.1",
    "asgi-correlation-id>=4.3.4",
    "asyncpg>=0.31.0",
    "dependency-injector>=4.48.3",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.1",
    "loguru>=0.7.3",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "pydantic-settings>=2.12.0",
    "python-json-logger>=4.0.0",
    "sqlmodel>=0.0.31",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "bcrypt==4.0.1",
    "langchain[google-genai]>=1.2.7",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=4.2.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary]>=3.2.1",
]

[tool.pyright]
typeCheckingMode = "off" # typechecking handled by mypy

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
lint = [
    "ruff>=0.14.14",
]
